refactor(text-to-image): moves "Client/SDXL/exports.ts" to "Client/SDXL/generateOutputFrom.ts"
