+++ src/features/TextToImage/Client/SDXL/generateOutputFrom.ts
