+++ src/features/TextToImage/Client/SDXL/index.ts
@@ -1,3 +1,3 @@
 export {
   TextToImage_Client_SDXL as default,
-} from './exports.ts';
+} from './generateOutputFrom.ts';
